Lab5/service-customer/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
flask-migrate==4.0.5
redis==5.0.1
marshmallow==3.20.1
python-dotenv==1.0.0 